# Conflicts:
#	src/main/java/net/kapitencraft/kap_lib/client/widget/BackButton.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/Checkbox.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/PositionedWidget.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/Widget.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/background/CutoutBackground.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/background/FillBackground.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/background/TextureBackground.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/menu/Menu.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/menu/drop_down/DropDownMenu.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/menu/drop_down/elements/BooleanElement.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/menu/drop_down/elements/Element.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/menu/drop_down/elements/ListElement.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/menu/scroll/ScrollableMenu.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/menu/scroll/elements/ScrollElement.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/select/CreateBannerWidget.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/select/MultiElementSelectorWidget.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/select/SelectColorWidget.java
#	src/main/java/net/kapitencraft/kap_lib/client/widget/text/MultiLineTextBox.java
#	src/main/java/net/kapitencraft/kap_lib/entity/fishing/AbstractFishingHook.java
#	src/main/java/net/kapitencraft/kap_lib/io/serialization/JsonSerializer.java
#	src/main/java/net/kapitencraft/kap_lib/io/serialization/NbtSerializer.java
#	src/main/java/net/kapitencraft/kap_lib/io/serialization/Serializer.java
#	src/main/java/net/kapitencraft/kap_lib/registry/ExtraLootItemConditions.java
#	src/main/java/net/kapitencraft/kap_lib/registry/ExtraLootItemFunctions.java
#	src/main/java/net/kapitencraft/kap_lib/registry/ExtraParticleTypes.java
#	src/main/java/net/kapitencraft/kap_lib/util/range/simple/DoubleNumberRange.java
#	src/main/java/net/kapitencraft/kap_lib/util/range/simple/NumberRange.java

+++ src/main/java/net/kapitencraft/kap_lib/util/DamageCounter.java
@@ -1,11 +1,11 @@
 package net.kapitencraft.kap_lib.util;
 
 public class DamageCounter {
-    private static double damage = 0;
+    private static float damage = 0;
     private static int attacked = 0;
     private static boolean shouldCount = false;
 
-    public static void increaseDamage(double damageToAdd) {
+    public static void increaseDamage(float damageToAdd) {
         if (shouldCount) {
             damage += damageToAdd;
             attacked++;
@@ -20,13 +20,13 @@ public class DamageCounter {
         DamageHolder holder = new DamageHolder(attacked, damage);
         if (reset) {
             attacked = 0;
-            damage = 0;
+                damage = 0;
             shouldCount = false;
         }
         return holder;
     }
 
-    public record DamageHolder(int hit, double damage) {
+    public record DamageHolder(int hit, float damage) {
 
         public boolean hasDamage() {
             return hit > 0 && damage > 0;
